Issue number: 10763 - hang (infinite loop) when having big min or max pool size Submitted by: Dies Koper Reviewed by:  Kenneth Saks
Changed populateCount's type to a long.


svn path=/trunk/; revision=35397


Former-commit-id: 109f3ca0c5082b87009192abf307c901c0d16d77

+++ v3/ejb/ejb-container/src/main/java/com/sun/ejb/containers/util/pool/NonBlockingPool.java
@@ -437,7 +437,7 @@ public class NonBlockingPool
             startTime = System.currentTimeMillis();
 
             ArrayList removeList = new ArrayList();
-            int populateCount = 0;
+            long populateCount = 0;
             synchronized (list) {
                 if ((inResizing == true) || poolClosed) {
                     return;
@@ -518,7 +518,7 @@ public class NonBlockingPool
                             + list.size() + "/" + maxPoolSize);
                 }
 
-                preload(populateCount);
+                preload((int)populateCount);
 
                 if(_logger.isLoggable(Level.FINE)) {
                     _logger.log(Level.FINE,
